requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
uvicorn==0.30.1
requests==2.32.3
response==0.5.0
pillow==10.3.0
pydantic==2.7.4
jsonpatch==1.33
jsonpointer==3.0.0
openai==0.28.0
httpx
Image
pillow
uuid
fastapi-cors
Upload